# openapi.yaml
openapi: 3.0.3
info:
  title: Oblixa route contract
  version: 0.1.0
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sb-access-token
    CronSecret:
      type: http
      scheme: bearer
    WebhookSignature:
      type: apiKey
      in: header
      name: stripe-signature
paths:
...
  /api/export/contracts/{jobId}:
    get:
      operationId: get_export_contracts_by_jobId
      summary: GET /api/export/contracts/{jobId}
      description: Generated route contract for src/app/api/export/contracts/[jobId]/route.ts.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "429":
          description: Rate limited
        "500":
          description: Unexpected server error
    post:
      operationId: post_export_contracts_by_jobId
      summary: POST /api/export/contracts/{jobId}
      description: Generated route contract for src/app/api/export/contracts/[jobId]/route.ts.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "409":
          description: Conflict
        "413":
          description: Payload too large
        "429":
          description: Rate limited
        "500":
          description: Unexpected server error
  /api/export/contracts:
    get:
      operationId: get_export_contracts
      summary: GET /api/export/contracts
      description: Generated route contract for src/app/api/export/contracts/route.ts.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "429":
          description: Rate limited
        "500":
          description: Unexpected server error
    post:
      operationId: post_export_contracts
      summary: POST /api/export/contracts
      description: Generated route contract for src/app/api/export/contracts/route.ts.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "409":
          description: Conflict
        "413":
          description: Payload too large
        "429":
          description: Rate limited
        "500":
          description: Unexpected server error
...